Divided huge file in 3 smaller ones.
